+++ src/main/java/com/github/f4b6a3/tsid/TsidCreator.java
@@ -24,19 +24,12 @@
 
 package com.github.f4b6a3.tsid;
 
-import com.github.f4b6a3.tsid.factory.TsidFactory;
-
 /**
- * A facade that generates time sortable IDs (TSIDs).
+ * A class that generates TSIDs (Time Sortable Identifier).
  * 
- * The term TSID stands for (roughly) Time Sortable ID. A TSID is a number that
- * is formed by the creation time followed by random bits.
+ * A TSID is formed by the creation time followed by random bits.
  */
 public final class TsidCreator {
-
-	public static final int NODE_LENGTH_256 = 8; // S
-	public static final int NODE_LENGTH_1024 = 10; // M
-	public static final int NODE_LENGTH_4096 = 12; // L
 
 	private TsidCreator() {
 	}
@@ -66,7 +59,7 @@ public final class TsidCreator {
 	 * @return a TSID
 	 */
 	public static Tsid getTsid256() {
-		return TsidFactory256Holder.INSTANCE.create();
+		return Factory256Holder.INSTANCE.create();
 	}
 
 	/**
@@ -94,7 +87,7 @@ public final class TsidCreator {
 	 * @return a TSID
 	 */
 	public static Tsid getTsid1024() {
-		return TsidFactory1024Holder.INSTANCE.create();
+		return Factory1024Holder.INSTANCE.create();
 	}
 
 	/**
@@ -122,99 +115,18 @@ public final class TsidCreator {
 	 * @return a TSID number
 	 */
 	public static Tsid getTsid4096() {
-		return TsidFactory4096Holder.INSTANCE.create();
+		return Factory4096Holder.INSTANCE.create();
 	}
 
-	/**
-	 * Returns a {@link TsidFactory}.
-	 * 
-	 * The node identifier is defined by parameter. A null argument means RANDOM
-	 * node identifier.
-	 * 
-	 * It supports up to 256 nodes.
-	 * 
-	 * It can generate up to 16,384 TSIDs per millisecond per node.
-	 * 
-	 * ## Random component settings:
-	 * 
-	 * - Node bit length: 8
-	 * 
-	 * - Counter bit length: 14
-	 * 
-	 * - Maximum node: 256 (2^8)
-	 * 
-	 * - Maximum counter: 16,384 (2^14)
-	 * 
-	 * @param node a node identifier
-	 * @return a {@link TsidFactory}
-	 */
-	public static TsidFactory getTsidFactory256(Integer node) {
-		return new TsidFactory(node, NODE_LENGTH_256);
+	private static class Factory256Holder {
+		static final TsidFactory INSTANCE = TsidFactory.newInstance256();
 	}
 
-	/**
-	 * Returns a {@link TsidFactory}.
-	 * 
-	 * The node identifier is defined by parameter. A null argument means RANDOM
-	 * node identifier.
-	 * 
-	 * It supports up to 1,024 nodes.
-	 * 
-	 * It can generate up to 4,096 TSIDs per millisecond per node.
-	 * 
-	 * ## Random component settings:
-	 * 
-	 * - Node bit length: 10
-	 * 
-	 * - Counter bit length: 12
-	 * 
-	 * - Maximum node: 1,024 (2^10)
-	 * 
-	 * - Maximum counter: 4,096 (2^12)
-	 * 
-	 * @param node a node identifier
-	 * @return a {@link TsidFactory}
-	 */
-	public static TsidFactory getTsidFactory1024(Integer node) {
-		return new TsidFactory(node, NODE_LENGTH_1024);
+	private static class Factory1024Holder {
+		static final TsidFactory INSTANCE = TsidFactory.newInstance1024();
 	}
 
-	/**
-	 * Returns a {@link TsidFactory}.
-	 * 
-	 * The node identifier is defined by parameter. A null argument means RANDOM
-	 * node identifier.
-	 * 
-	 * It supports up to 4,096 nodes.
-	 * 
-	 * It can generate up to 1,024 TSIDs per millisecond per node.
-	 * 
-	 * ## Random component settings:
-	 * 
-	 * - Node bit length: 12
-	 * 
-	 * - Counter bit length: 10
-	 * 
-	 * - Maximum node: 4,096 (2^12)
-	 * 
-	 * - Maximum counter: 1,024 (2^10)
-	 * 
-	 * @param node a node identifier
-	 * @return a {@link TsidFactory}
-	 */
-	public static TsidFactory getTsidFactory4096(Integer node) {
-		return new TsidFactory(node, NODE_LENGTH_4096);
-	}
-
-	private static class TsidFactory256Holder {
-		static final TsidFactory INSTANCE = getTsidFactory256(null);
-	}
-
-	private static class TsidFactory1024Holder {
-		static final TsidFactory INSTANCE = getTsidFactory1024(null);
-	}
-
-	private static class TsidFactory4096Holder {
-		static final TsidFactory INSTANCE = getTsidFactory4096(null);
+	private static class Factory4096Holder {
+		static final TsidFactory INSTANCE = TsidFactory.newInstance4096();
 	}
 }