+++ src/test/java/com/github/f4b6a3/tsid/TestSuite.java
@@ -3,14 +3,13 @@ package com.github.f4b6a3.tsid;
 import org.junit.runner.RunWith;
 import org.junit.runners.Suite;
 
-import com.github.f4b6a3.tsid.util.TsidTimeTest;
-import com.github.f4b6a3.tsid.util.internal.SettingsUtilTest;
 import com.github.f4b6a3.tsid.factory.TsidFactory00001Test;
 import com.github.f4b6a3.tsid.factory.TsidFactory00064Test;
 import com.github.f4b6a3.tsid.factory.TsidFactory00256Test;
 import com.github.f4b6a3.tsid.factory.TsidFactory01024Test;
 import com.github.f4b6a3.tsid.factory.TsidFactory04096Test;
 import com.github.f4b6a3.tsid.factory.TsidFactory16384Test;
+import com.github.f4b6a3.tsid.internal.SettingsUtilTest;
 
 @RunWith(Suite.class)
 @Suite.SuiteClasses({
@@ -20,7 +19,6 @@ import com.github.f4b6a3.tsid.factory.TsidFactory16384Test;
 	TsidFactory01024Test.class,
 	TsidFactory04096Test.class,
 	TsidFactory16384Test.class,
-	TsidTimeTest.class,
 	TsidTest.class,
 	SettingsUtilTest.class,
 })