+++ src/test/java/com/github/f4b6a3/tsid/factory/TsidFactory16384Test.java
@@ -12,7 +12,7 @@ import java.util.Set;
 import org.junit.Test;
 
 import com.github.f4b6a3.tsid.Tsid;
-import com.github.f4b6a3.tsid.factory.TsidFactory;
+import com.github.f4b6a3.tsid.TsidFactory;
 
 public class TsidFactory16384Test {
 
@@ -33,11 +33,11 @@ public class TsidFactory16384Test {
 
 		long startTime = System.currentTimeMillis();
 
-		TsidFactory creator = new TsidFactory(null, NODE_LENGTH);
+		TsidFactory factory = TsidFactory.builder().withNodeBitLength(NODE_LENGTH).withRandom(random).build();
 
 		long[] list = new long[COUNTER_MAX];
 		for (int i = 0; i < COUNTER_MAX; i++) {
-			list[i] = creator.create().toLong();
+			list[i] = factory.create().toLong();
 		}
 
 		long endTime = System.currentTimeMillis();
@@ -53,11 +53,12 @@ public class TsidFactory16384Test {
 		long startTime = System.currentTimeMillis();
 
 		int node = random.nextInt();
-		TsidFactory creator = new TsidFactory(node, NODE_LENGTH);
+		TsidFactory factory = TsidFactory.builder().withNode(node).withNodeBitLength(NODE_LENGTH).withRandom(random)
+				.build();
 
 		long[] list = new long[COUNTER_MAX];
 		for (int i = 0; i < COUNTER_MAX; i++) {
-			list[i] = creator.create().toLong();
+			list[i] = factory.create().toLong();
 		}
 
 		long endTime = System.currentTimeMillis();
@@ -72,11 +73,11 @@ public class TsidFactory16384Test {
 
 		long startTime = System.currentTimeMillis();
 
-		TsidFactory creator = new TsidFactory(null, NODE_LENGTH);
+		TsidFactory factory = TsidFactory.builder().withNodeBitLength(NODE_LENGTH).withRandom(random).build();
 
 		String[] list = new String[COUNTER_MAX];
 		for (int i = 0; i < COUNTER_MAX; i++) {
-			list[i] = creator.create().toString();
+			list[i] = factory.create().toString();
 		}
 
 		long endTime = System.currentTimeMillis();
@@ -92,11 +93,12 @@ public class TsidFactory16384Test {
 		long startTime = System.currentTimeMillis();
 
 		int node = random.nextInt();
-		TsidFactory creator = new TsidFactory(node, NODE_LENGTH);
+		TsidFactory factory = TsidFactory.builder().withNode(node).withNodeBitLength(NODE_LENGTH).withRandom(random)
+				.build();
 
 		String[] list = new String[COUNTER_MAX];
 		for (int i = 0; i < COUNTER_MAX; i++) {
-			list[i] = creator.create().toString();
+			list[i] = factory.create().toString();
 		}
 
 		long endTime = System.currentTimeMillis();
@@ -113,11 +115,11 @@ public class TsidFactory16384Test {
 		Thread[] threads = new Thread[THREAD_TOTAL];
 		int counterMax = COUNTER_MAX / THREAD_TOTAL;
 
-		TsidFactory sharedCreator = new TsidFactory(null, NODE_LENGTH);
-
 		// Instantiate and start many threads
 		for (int i = 0; i < THREAD_TOTAL; i++) {
-			threads[i] = new TestThread(sharedCreator, counterMax);
+			TsidFactory factory = TsidFactory.builder().withNode(i).withNodeBitLength(NODE_LENGTH).withRandom(random)
+					.build();
+			threads[i] = new TestThread(factory, counterMax);
 			threads[i].start();
 		}